# docs/openapi/stigmem.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SubscriptionCreateRequest": {
        "properties": {
          "delivery_address": {
            "description": "Webhook URL (on_change=webhook) or identity URI (on_change=wake)",
            "minLength": 1,
            "title": "Delivery Address",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client idempotency key",
            "title": "Idempotency Key"
          },
          "on_change": {
            "description": "'webhook' or 'wake'",
            "title": "On Change",
            "type": "string"
          },
          "target": {
            "description": "Scope name (e.g. 'local') or entity URI to watch",
            "minLength": 1,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "on_change",
          "delivery_address"
        ],
        "title": "SubscriptionCreateRequest",
        "type": "object"
      },
      "SubscriptionEventRecord": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "delivered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "delivery_attempts": {
            "title": "Delivery Attempts",
            "type": "integer"
          },
          "delivery_status": {
            "title": "Delivery Status",
            "type": "string"
          },
          "entity_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Uri"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "fact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "subscription_id": {
            "title": "Subscription Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subscription_id",
          "event_type",
          "entity_uri",
          "fact_id",
          "payload",
          "created_at",
          "delivered_at",
          "delivery_status",
          "delivery_attempts"
        ],
        "title": "SubscriptionEventRecord",
        "type": "object"
      },
      "SubscriptionEventsResponse": {
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/SubscriptionEventRecord"
            },
            "title": "Events",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "total",
          "cursor"
        ],
        "title": "SubscriptionEventsResponse",
        "type": "object"
      },
      "SubscriptionListResponse": {
        "properties": {
          "subscriptions": {
            "items": {
              "$ref": "#/components/schemas/SubscriptionRecord"
            },
            "title": "Subscriptions",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "subscriptions",
          "total"
        ],
        "title": "SubscriptionListResponse",
        "type": "object"
      },
      "SubscriptionRecord": {
        "properties": {
          "circuit_open": {
            "title": "Circuit Open",
            "type": "boolean"
          },
          "consecutive_failures": {
            "title": "Consecutive Failures",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "delivery_address": {
            "title": "Delivery Address",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "last_delivered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivered At"
          },
          "on_change": {
            "title": "On Change",
            "type": "string"
          },
          "subscriber_identity": {
            "title": "Subscriber Identity",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "target_kind": {
            "title": "Target Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subscriber_identity",
          "target",
          "target_kind",
          "on_change",
          "delivery_address",
          "idempotency_key",
          "created_at",
          "last_delivered_at",
          "circuit_open",
          "consecutive_failures"
        ],
        "title": "SubscriptionRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Reference node implementing the Stigmem v0.9.0a1 HTTP API \u2014 facts, federation, gardens, recall, subscriptions, audit, identity, content-addressed fact IDs. Cross-cutting features (tombstones, time-travel, multi-tenant) are opt-in plugins.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Stigmem Reference Node",
    "version": "0.9.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/subscriptions": {
      "get": {
        "description": "List the authenticated caller's subscriptions (BOLA: own only).",
        "operationId": "list_subscriptions_v1_subscriptions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Subscriptions",
        "tags": [
          "subscriptions"
        ]
      },
      "post": {
        "description": "Create a subscription (Spec-X7-Subscriptions).\n\nsubscriber_identity is always the caller (BOLA).",
        "operationId": "create_subscription_v1_subscriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Subscription",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/v1/subscriptions/{subscription_id}": {
      "delete": {
        "description": "Cancel a subscription (BOLA: only the owner may delete).",
        "operationId": "delete_subscription_v1_subscriptions__subscription_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "subscription_id",
            "required": true,
            "schema": {
              "title": "Subscription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Subscription",
        "tags": [
          "subscriptions"
        ]
      },
      "get": {
        "description": "Get a subscription by ID (BOLA: returns 404 if caller does not own it).",
        "operationId": "get_subscription_v1_subscriptions__subscription_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "subscription_id",
            "required": true,
            "schema": {
              "title": "Subscription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Subscription",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/v1/subscriptions/{subscription_id}/events": {
      "get": {
        "description": "Replay window: return delivery events for a subscription (Spec-X7-Subscriptions).\n\nResults are bounded to the configured replay window (default 24 h).",
        "operationId": "list_subscription_events_v1_subscriptions__subscription_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "subscription_id",
            "required": true,
            "schema": {
              "title": "Subscription Id",
              "type": "string"
            }
          },
          {
            "description": "ISO 8601 timestamp; return events created at or after this time",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 timestamp; return events created at or after this time",
              "title": "Since"
            }
          },
          {
            "description": "Opaque pagination cursor (event id)",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor (event id)",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Subscription Events",
        "tags": [
          "subscriptions"
        ]
      }
    },
...
  }
}